+++ bin-src/tasks/verify.js
@@ -2,9 +2,11 @@ import { createTask, transitionTo } from '../lib/tasks';
 import listingStories from '../ui/messages/info/listingStories';
 import storybookPublished from '../ui/messages/info/storybookPublished';
 import buildLimited from '../ui/messages/warnings/buildLimited';
+import noAncestorBuild from '../ui/messages/warnings/noAncestorBuild';
 import paymentRequired from '../ui/messages/warnings/paymentRequired';
 import snapshotQuotaReached from '../ui/messages/warnings/snapshotQuotaReached';
 import { initial, dryRun, pending, runOnly, runOnlyFiles, success } from '../ui/tasks/verify';
+import turboSnapEnabled from '../ui/messages/info/turboSnapEnabled';
 
 const TesterCreateBuildMutation = `
   mutation TesterCreateBuildMutation($input: CreateBuildInput!, $isolatorUrl: String!) {
@@ -16,6 +18,7 @@ const TesterCreateBuildMutation = `
       testCount
       actualTestCount: testCount(statuses: [IN_PROGRESS])
       actualCaptureCount
+      inheritedCaptureCount
       webUrl
       cachedUrl
       reportToken
@@ -72,10 +75,10 @@ export const setEnvironment = async (ctx) => {
 };
 
 export const createBuild = async (ctx, task) => {
-  const { client, git, log, isolatorUrl, options, onlyStoryFiles } = ctx;
-  const { list, only, patchBaseRef, patchHeadRef, preserveMissingSpecs } = options;
-  const { version, matchesBranch, changedFiles, ...commitInfo } = git; // omit some fields
-  const autoAcceptChanges = matchesBranch(options.autoAcceptChanges);
+  const { list, only, patchBaseRef, patchHeadRef, preserveMissingSpecs } = ctx.options;
+  const { version, matchesBranch, changedFiles, ...commitInfo } = ctx.git; // omit some fields
+  const { isolatorUrl, rebuildForBuildId, onlyStoryFiles, turboSnap } = ctx;
+  const autoAcceptChanges = matchesBranch(ctx.options.autoAcceptChanges);
 
   // It's not possible to set both --only and --only-changed
   if (only) {
@@ -85,13 +88,18 @@ export const createBuild = async (ctx, task) => {
     transitionTo(runOnlyFiles)(ctx, task);
   }
 
-  const { createBuild: build } = await client.runQuery(
+  const { createBuild: build } = await ctx.client.runQuery(
     TesterCreateBuildMutation,
     {
       input: {
         ...commitInfo,
+        rebuildForBuildId,
         ...(only && { only }),
         ...(onlyStoryFiles && { onlyStoryFiles: Object.keys(onlyStoryFiles) }),
+        ...(turboSnap && { turboSnapEnabled: !turboSnap.bailReason }),
+        // GraphQL does not support union input types (yet), so we stringify the bailReason
+        // @see https://github.com/graphql/graphql-spec/issues/488
+        ...(turboSnap?.bailReason && { turboSnapBailReason: JSON.stringify(turboSnap.bailReason) }),
         autoAcceptChanges,
         cachedUrl: ctx.cachedUrl,
         environment: ctx.environment,
@@ -113,27 +121,35 @@ export const createBuild = async (ctx, task) => {
   ctx.isOnboarding = build.number === 1 || (build.autoAcceptChanges && !autoAcceptChanges);
 
   if (list) {
-    log.info(listingStories(build.tests));
+    ctx.log.info(listingStories(build.tests));
+  }
+
+  if (!ctx.isOnboarding && !ctx.git.parentCommits) {
+    ctx.log.warn(noAncestorBuild(ctx));
+  }
+
+  if (ctx.turboSnap && !ctx.turboSnap.bailReason) {
+    ctx.log.info(turboSnapEnabled(ctx));
   }
 
   if (build.wasLimited) {
     const { account } = build.app;
     if (account.exceededThreshold) {
-      log.warn(snapshotQuotaReached(account));
+      ctx.log.warn(snapshotQuotaReached(account));
       ctx.exitCode = 101;
     } else if (account.paymentRequired) {
-      log.warn(paymentRequired(account));
+      ctx.log.warn(paymentRequired(account));
       ctx.exitCode = 102;
     } else {
       // Future proofing for reasons we aren't aware of
-      log.warn(buildLimited(account));
+      ctx.log.warn(buildLimited(account));
       ctx.exitCode = 100;
     }
   }
 
   transitionTo(success, true)(ctx, task);
 
-  if (list || ctx.isPublishOnly || matchesBranch(options.exitOnceUploaded)) {
+  if (list || ctx.isPublishOnly || matchesBranch(ctx.options.exitOnceUploaded)) {
     ctx.exitCode = 0;
     ctx.skipSnapshots = true;
     ctx.log.info(storybookPublished(ctx));